Рефакторинг (разбивает main на два компонента + драит код)
